+++ src/main/java/com/daruda/darudaserver/domain/user/service/UserService.java
@@ -110,20 +110,6 @@ public class UserService {
 
     }
 
-    private void validateRefreshToken(final String refreshToken){
-        JwtValidationType jwtValidationType =jwtTokenProvider.validateToken(refreshToken);
-
-        if(!jwtValidationType.equals(JwtValidationType.VALID_JWT)){
-            throw switch(jwtValidationType){
-                case EXPIRED_JWT_TOKEN -> new UnauhtorizedException(ErrorCode.REFRESH_TOKEN_EXPIRED_ERROR);
-                case INVALID_JWT_TOKEN -> new BadRequestException(ErrorCode.INVALID_REFRESH_TOKEN_ERROR);
-                case INVALID_JWT_SIGNATURE -> new BadRequestException(ErrorCode.REFRESH_TOKEN_SIGNATURE_ERROR);
-                case UNSUPPORTED_JWT_TOKEN -> new BadRequestException(ErrorCode.UNSUPPORTED_REFRESH_TOKEN_ERROR);
-                case EMPTY_JWT -> new BadRequestException(ErrorCode.REFREH_TOKEN_EMPTY_ERROR);
-                default -> new BusinessException(ErrorCode.UNSUPPORTED_REFRESH_TOKEN_ERROR);
-            };
-        }
-    }
 
     private void verifyUserIdWithStoredToken(final Long userId, final String refreshToken){
         Long storedUserId = tokenService.findIdByRefreshToken(refreshToken);
